Support up to 500 laps: multiplayer slider, lazy lap audio, new clips (#121)
* Support up to 500 laps (room/race path)

Widen the lap count from byte (max 255) to int across the room, race, and
track-package paths so single race, time trial, and multiplayer rooms can be
set up to 500 laps. The byte cap was only an artifact of the old 16-lap limit,
not a deliberate design constraint.

Changes are a near 1:1 type widening:
  - byte -> int on every Laps/NrOfLaps storage and wire field (packets,
    GameRoom, TrackData, room snapshot/event, track-package manifest).
  - Serialization read/write switched to Int32 with matching payload sizes
    (client and server sides kept symmetric).
  - Lap limits/validation raised 16 -> 500 (session MaxLaps consts, room
    option list, settings clamp, server validation, menu range).

Protocol shape tests updated for the wider Laps field and the packet-shape
snapshot regenerated (RoomEvent Laps now 4 bytes). Full test suite passes
(426/426). Cross-build note: this changes the wire format, so new and old
builds are not interoperable on lap-bearing packets.


* Support up to 500 laps: multiplayer slider, lazy lap audio, new clips

Multiplayer room options now use the same Slider widget as local race
settings for "Number of laps", so HOME/END/PAGE UP/PAGE DOWN big-step
shortcuts work there too (previously a RadioButton with single-step
LEFT/RIGHT and wraparound). Adds the raw-value accessor and drops the
now-unused LapCountOptions array.

Lap "laps to go" sounds are now loaded lazily on demand instead of
bulk-decoded at race construction, across single, time trial, and
multiplayer sessions. The Progress subsystems take a Func<int,Source?>
accessor backed by a nullable slot array; each laps2go clip is decoded
and cached only when that lap is reached. This removes a start-of-race
decode hitch and caps resident memory to laps actually played through,
which matters now that races can be up to 500 laps. Matches the
existing lazy pattern used for number/position/finish sounds.

Adds laps2go16..499 English audio clips; non-English packs fall back
to English for any clip they have not yet translated.


* Convert laps2go300 / laps2go430 from mp3 to ogg

These two lap-countdown clips were committed as .mp3 while every other
laps2go clip is .ogg. The language-sound resolver only appends .ogg
(AssetPaths.ResolveLanguageSoundPath), so at 300 and 430 laps the callout
failed to resolve and fell through to error.wav.

Transcoded with ffmpeg to match Paul's other clips exactly: libvorbis
-q:a 4 (nominal 86 kbps), 44100 Hz, mono. Durations preserved.


---------

+++ top_speed_net/TopSpeed.Shared/Data/Tracks/Types.cs
@@ -164,7 +164,7 @@ namespace TopSpeed.Data
         public IReadOnlyDictionary<string, TrackSoundSourceDefinition> SoundSources { get; }
         public string? SourcePath { get; }
         public int Length => Definitions.Length;
-        public byte Laps { get; set; }
+        public int Laps { get; set; }
 
         /// <summary>
         /// Whether this track has a pit area. True unless the <c>[meta]</c> section explicitly sets
@@ -196,7 +196,7 @@ namespace TopSpeed.Data
             TrackWeather weather,
             TrackAmbience ambience,
             TrackDefinition[] definitions,
-            byte laps = 0,
+            int laps = 0,
             string? name = null,
             string? version = null,
             IReadOnlyDictionary<string, string>? metadata = null,
@@ -228,7 +228,7 @@ namespace TopSpeed.Data
             IReadOnlyDictionary<string, TrackWeatherProfile>? weatherProfiles,
             TrackAmbience ambience,
             TrackDefinition[] definitions,
-            byte laps = 0,
+            int laps = 0,
             string? name = null,
             string? version = null,
             IReadOnlyDictionary<string, string>? metadata = null,
@@ -276,7 +276,7 @@ namespace TopSpeed.Data
                 : TrackWeatherProfile.CreatePreset(TrackWeatherProfile.DefaultProfileId, TrackWeather.Sunny);
         }
 
-        public TrackData WithLaps(byte laps)
+        public TrackData WithLaps(int laps)
         {
             return new TrackData(
                 UserDefined,
